fix: fix validation
