+++ my-forms/src/Components/Pages/ReactHookFormPage/index.tsx
@@ -1,5 +1,6 @@
 import { FC } from 'react';
 import { Link } from 'react-router-dom';
+
 import { ROUTES } from '../../../utils/constants/routes';
 
 const ReactHookFormPage: FC = () => {
