+++ my-forms/src/store/reducers/dataSlice.ts
@@ -1,4 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
+
 import { initialStateData } from './slice.type';
 
 export const dataSlice = createSlice({
